@@ -7,7 +7,7 @@
             <MudCardHeader>
                 <MudCardHeaderContent>
                     <MudText>Register Id: @HardwareService.ClientIdentifier</MudText>
-                    <MudText>Register Address: @HardwareService.IpAddress</MudText>
+                    <MudText>Register Address: @HardwareService.IpAddress (@HardwareService.HostName)</MudText>
                 </MudCardHeaderContent>
             </MudCardHeader>
             <MudCardContent>
