@@ -1,14 +1,13 @@
 ﻿// <auto-generated />
-using System;
+
+#nullable disable
+
 using Innkeep.Db.Server.Context;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Migrations;
-using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
-#nullable disable
-
-namespace Innkeep.Server.Db.Migrations
+namespace Innkeep.Db.Server.Migrations.Server
 {
     [DbContext(typeof(InnkeepServerContext))]
     [Migration("20240621101013_Created Fiskaly TSE Specific Config")]
