@@ -1,13 +1,13 @@
 ﻿// <auto-generated />
+
+#nullable disable
+
 using Innkeep.Db.Server.Context;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Migrations;
-using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
-#nullable disable
-
-namespace Innkeep.Db.Server.Migrations
+namespace Innkeep.Db.Server.Migrations.Server
 {
     [DbContext(typeof(InnkeepServerContext))]
     [Migration("20240816101742_Added Checkin List Config")]
