@@ -1,17 +1,18 @@
 ﻿// <auto-generated />
-
-#nullable disable
-
+using System;
 using Innkeep.Db.Server.Context;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Migrations;
+using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
+
+#nullable disable
 
 namespace Innkeep.Db.Server.Migrations.Transaction
 {
     [DbContext(typeof(InnkeepTransactionContext))]
-    [Migration("20240817122756_Initial")]
-    partial class Initial
+    [Migration("20240817141408_Added Register Id")]
+    partial class AddedRegisterId
     {
         /// <inheritdoc />
         protected override void BuildTargetModel(ModelBuilder modelBuilder)
@@ -37,14 +38,17 @@ namespace Innkeep.Db.Server.Migrations.Transaction
 
                     b.Property<string>("ClientId")
                         .IsRequired()
+                        .HasMaxLength(40)
                         .HasColumnType("TEXT");
 
                     b.Property<string>("EventId")
                         .IsRequired()
+                        .HasMaxLength(40)
                         .HasColumnType("TEXT");
 
                     b.Property<string>("OrderSecret")
                         .IsRequired()
+                        .HasMaxLength(40)
                         .HasColumnType("TEXT");
 
                     b.Property<string>("ReceiptJson")
@@ -53,6 +57,12 @@ namespace Innkeep.Db.Server.Migrations.Transaction
 
                     b.Property<string>("ReceiptType")
                         .IsRequired()
+                        .HasMaxLength(32)
+                        .HasColumnType("TEXT");
+
+                    b.Property<string>("RegisterId")
+                        .IsRequired()
+                        .HasMaxLength(40)
                         .HasColumnType("TEXT");
 
                     b.Property<DateTime>("TransactionDate")
@@ -60,6 +70,7 @@ namespace Innkeep.Db.Server.Migrations.Transaction
 
                     b.Property<string>("TssId")
                         .IsRequired()
+                        .HasMaxLength(40)
                         .HasColumnType("TEXT");
 
                     b.HasKey("Id");
