@@ -1,4 +1,5 @@
-﻿using System.ComponentModel.DataAnnotations.Schema;
+﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using Innkeep.Db.Models;
 
 namespace Innkeep.Db.Server.Models.Transaction;
@@ -7,15 +8,24 @@ public class TransactionModel : AbstractDbItem
 {
 	public DateTime TransactionDate { get; set; }
 
+	[MaxLength(32)]
 	public string ReceiptType { get; set; } = "";
 	
+	[MaxLength(40)]
 	public string TssId { get; set; } = "";
 	
+	[MaxLength(40)]
 	public string ClientId { get; set; } = "";
 	
+	[MaxLength(40)]
 	public string EventId { get; set; } = "";
 
+	[MaxLength(40)]
 	public string OrderSecret { get; set; } = "";
+
+	[MaxLength(40)]
+	public string RegisterId { get; set; } = "";
+	
 	public decimal AmountRequested { get; set; }
 	public decimal AmountGiven { get; set; }
 	
@@ -24,5 +34,5 @@ public class TransactionModel : AbstractDbItem
 	public string ReceiptJson { get; set; } = "";
 
 	[NotMapped]
-	public decimal TotalChange => AmountGiven - AmountRequested;
+	public decimal TotalChange => AmountGiven - AmountBack;
 }
