@@ -1,5 +1,8 @@
 ﻿<MudNavMenu>
-    <MudNavGroup Title="Config" Expanded="true">
+    <MudNavGroup Title="Register" Expanded="true">
+        <MudNavLink Href="/register/dashboard">Dashboard</MudNavLink>
+    </MudNavGroup>
+    <MudNavGroup Title="Config" Expanded="false">
         <MudNavGroup Title="Pretix" Expanded="true">
             <MudNavLink Href="/config/pretix">Token</MudNavLink>
             <MudNavLink Href="/config/event">Event & Organizer</MudNavLink>
