@@ -1,4 +1,7 @@
-﻿using System.Windows;
+﻿using System.Net;
+using System.Net.Security;
+using System.Security.Cryptography.X509Certificates;
+using System.Windows;
 using Innkeep.Db.Server.Context;
 using Innkeep.Resources;
 using Innkeep.Server.Startup;
@@ -17,7 +20,8 @@ public partial class App : Application
 	{
 		ThreadCultureHelper.SetInvariant();
 		base.OnStartup(e);
-
+		
+		ServicePointManager.ServerCertificateValidationCallback = LocalHostTesting;
 		LoggingManager.InitializeLogger("Innkeep Server");
 
 		var host = KestrelBuilder.Build();
@@ -30,5 +34,15 @@ public partial class App : Application
 
 		var mainWindow = new MainWindow(host);
 		mainWindow.Show();
+	}
+	
+	private static bool LocalHostTesting(
+		object sender,
+		X509Certificate? certificate,
+		X509Chain? chain,
+		SslPolicyErrors sslPolicyErrors
+	)
+	{
+		return true;
 	}
 }
