@@ -1,4 +1,5 @@
 ﻿using Innkeep.Api.Enum.Fiskaly.Transaction;
+using Innkeep.Api.Enum.Shared;
 using Innkeep.Api.Models.Fiskaly.Objects.Transaction;
 using Innkeep.Api.Models.Fiskaly.Request.Transaction;
 using Innkeep.Api.Models.Internal;
@@ -8,7 +9,7 @@ namespace Innkeep.Services.Server.Fiskaly;
 
 public partial class FiskalyTransactionService
 {
-	public async Task<TransferReceipt?> CompleteTransferTransaction()
+	public async Task<TransferReceipt> CompleteTransferTransaction(ClientTransfer model)
 	{
 		var request = new FiskalyTransactionUpdateRequest()
 		{
@@ -22,9 +23,9 @@ public partial class FiskalyTransactionService
 				{
 					Receipt = new FiskalyReceipt()
 					{
-						ReceiptType = ReceiptType.Receipt,
-						AmountsPerVatRate = [],//TransferVatRates(transaction),
-						AmountsPerPaymentType = [],//TransferPaymentTypes(transaction),
+						ReceiptType = ReceiptType.Transfer,
+						AmountsPerVatRate = TransferVatRates(model.Amount * model.Factor),
+						AmountsPerPaymentType = TransferPaymentTypes(model.Amount * model.Factor),
 					},
 				},
 			},
@@ -32,7 +33,43 @@ public partial class FiskalyTransactionService
 		};
 
 		var result = await transactionRepository.UpdateTransaction(request);
+		
+		return new TransferReceipt
+		{
+			Amount = model.Amount,
+			IsRetrieve = model.IsRetrieve,
+			BookingTime = DateTime.Now,
+			TransactionCounter = result?.Number ?? -1,
+			QrCode = result?.QrCodeData ?? "TSS ERROR",
+		};
+	}
 
-		return new();
+	private static List<FiskalyAmountPerVatRate> TransferVatRates(decimal amount)
+	{
+		var list = new List<FiskalyAmountPerVatRate>
+		{
+			new ()
+			{
+				Amount = amount,
+				VatRate = VatRate.Null,
+			},
+		};
+
+		return list;
+	}
+
+	private static List<FiskalyAmountPerPaymentType> TransferPaymentTypes(decimal amount)
+	{
+		var list = new List<FiskalyAmountPerPaymentType>
+		{
+			new()
+			{
+				Amount = amount,
+				CurrencyCode = CurrencyCode.EUR,
+				PaymentType = PaymentType.Cash,
+			},
+		};
+
+		return list;
 	}
 }
